+++ packages/db/NPostgres.test.ts
@@ -1,3 +1,4 @@
+import { genEvent, jsonlEvents } from '@nostrify/nostrify/test';
 import { NostrEvent, NostrFilter } from '@nostrify/types';
 import { assert, assertEquals, assertRejects } from '@std/assert';
 import { Kysely, LogConfig, LogEvent } from 'kysely';
@@ -56,25 +57,6 @@ async function withoutDebug(callback: () => Promise<void>) {
   if (typeof DEBUG === 'string') {
     Deno.env.set('DEBUG', DEBUG);
   }
-}
-
-/** Import a JSONL fixture by name in tests. */
-export async function jsonlEvents(path: string): Promise<NostrEvent[]> {
-  const data = await Deno.readTextFile(path);
-  return data.split('\n').map((line) => JSON.parse(line));
-}
-
-/** Generate an event for use in tests. */
-export function genEvent(t: Partial<NostrEvent> = {}, sk: Uint8Array = generateSecretKey()): NostrEvent {
-  const { id, kind, pubkey, tags, content, created_at, sig } = finalizeEvent({
-    kind: 255,
-    created_at: 0,
-    content: '',
-    tags: [],
-    ...t,
-  }, sk);
-
-  return { id, kind, pubkey, tags, content, created_at, sig };
 }
 
 Deno.test('NPostgres.migrate', { ignore: !databaseUrl }, async () => {