+++ packages/nostrify/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added ReplyBotPolicy to block replies from bots in the same second.
+
 ## 0.30.1 - 2024-08-27
 
 ### Fixed